backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.4",
    "celery[redis]>=5.6.2",
    "chromadb>=1.5.5",
    "cryptography>=46.0.5",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.133.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.48",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.3.3",
]
